architectural refactoring added builder for Node
