--- id3/src/dataset/data/tree/Node.scala
@@ -1,5 +1,0 @@
-package dataset.data.tree
-
-case class Node[A: Ordering, B](attribute: B,
-                                nodes: List[NodeConnection[A, B]] = List.empty,
-                                leafs: List[LeafConnection[A]] = List.empty)
