+++ id3/src/id3/algorithm/Id3Classifier.scala
@@ -1,6 +1,6 @@
-package id3
+package id3.algorithm
 
-import dataset.data.tree.Node
+import id3.data.tree.Node
 
 object Id3Classifier {
   def classify[A: Ordering, B](id3Tree: Node[A, B], toClassify: Map[String, String]): Option[String] = {
