+++ id3/src/id3/data/BestAttribute.scala
@@ -1,4 +1,4 @@
-package dataset.data
+package id3.data
 
 case class BestAttribute[A: Ordering, B](attribute: B,
                                          subsets: List[Subset[A, B]])
