+++ id3/src/id3/data/Dataset.scala
@@ -1,4 +1,4 @@
-package dataset.data
+package id3.data
 
 case class Dataset[A: Ordering, B](attribute: B,
                                    data: List[A])
