+++ id3/src/id3/data/tree/Leaf.scala
@@ -1,4 +1,4 @@
-package dataset.data.tree
+package id3.data.tree
 
 case class Leaf[A: Ordering](attribute: A)
