+++ id3/src/id3/data/tree/NodeConnection.scala
@@ -1,3 +1,3 @@
-package dataset.data.tree
+package id3.data.tree
 
 case class NodeConnection[A: Ordering, B](arc: A, to: Node[A, B], probability: Double = 1.0)
